+++ src/main/java/com/biblioteca/gerenciamento/service/TurmaService.java
@@ -8,16 +8,16 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 @Service
 public class TurmaService {
+
     @Autowired
     private TurmaRepository turmaRepository;
 
     public Optional<Turma> findById(int id) {
-        return ResponseEntity.ok(turmaRepository.findById(id)).getBody();
+        return turmaRepository.findById(id);
     }
 
     public ResponseEntity<Turma> create(Turma turma) {
@@ -25,37 +25,30 @@ public class TurmaService {
     }
 
     public ResponseEntity<Turma> updateParcial(int id, TurmaUpdateDto dto) {
-
         Optional<Turma> optional = turmaRepository.findById(id);
-
         if (optional.isEmpty()) {
             return ResponseEntity.notFound().build();
         }
-
         Turma turma = optional.get();
-
         if (dto.getAnoLetivo() != null) {
             turma.setAnoLetivo(dto.getAnoLetivo());
         }
-
         if (dto.getSerie() != null) {
             turma.setSerie(dto.getSerie());
         }
-
         if (dto.getIdCurso() != null) {
             turma.setIdCurso(dto.getIdCurso());
         }
-
         if (dto.getIdCampus() != null) {
             turma.setIdCampus(dto.getIdCampus());
         }
-
-        Turma updated = turmaRepository.save(turma);
-
-        return ResponseEntity.ok(updated);
+        return ResponseEntity.ok(turmaRepository.save(turma));
     }
 
     public ResponseEntity<Void> delete(Integer id) {
+        if (!turmaRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         turmaRepository.deleteById(id);
         return ResponseEntity.ok().build();
     }
